requirements.txt
# default Django settings
asgiref==3.4.1
Django==3.2.5
pytz==2021.1
sqlparse==0.4.1

# auth
djangorestframework
djangorestframework_simplejwt
drf-yasg
# debug
django-debug-toolbar

# environment
python-dotenv

# automated bot
aiohttp